feat: saprated client side components for better randring
